Moved plot_1d and plot_2d into the plotting module, akin to hist
