+++ anesthetic/plotting/__init__.py
@@ -1,2 +1,4 @@
 """Plotting public API."""
-from anesthetic.plotting._core import PlotAccessor # noqa: 401
+from anesthetic.plotting._core import ( # noqa: 401
+    PlotAccessor, plot_1d, plot_2d
+)
