+++ anesthetic/weighted_pandas.py
@@ -7,7 +7,7 @@ from numpy.ma import masked_array
 from anesthetic.utils import (compress_weights, channel_capacity, quantile,
                               temporary_seed)
 from pandas.core.accessor import CachedAccessor
-from anesthetic.plotting import PlotAccessor
+import anesthetic.plotting
 
 
 class _WeightedObject(object):
@@ -59,7 +59,7 @@ class _WeightedObject(object):
         """Effective number of samples."""
         return channel_capacity(self.weights)
 
-    plot = CachedAccessor("plot", PlotAccessor)
+    plot = CachedAccessor("plot", anesthetic.plotting.PlotAccessor)
 
 
 class WeightedSeries(_WeightedObject, Series):
@@ -300,6 +300,9 @@ class WeightedDataFrame(_WeightedObject, DataFrame):
         df = DataFrame(data=data, index=index, columns=self.columns)
         df.index = df.index.get_level_values('#')
         return df
+
+    plot_1d = anesthetic.plotting.plot_1d
+    plot_2d = anesthetic.plotting.plot_2d
 
     @property
     def _constructor_sliced(self):
